properties relocation/merge
