+++ src/main/java/be/uantwerpen/minelabs/block/entity/BohrBlockEntity.java
@@ -4,7 +4,9 @@ import be.uantwerpen.minelabs.block.BohrBlock;
 import be.uantwerpen.minelabs.block.BohrPart;
 import be.uantwerpen.minelabs.inventory.ImplementedInventory;
 import be.uantwerpen.minelabs.item.Items;
-import be.uantwerpen.minelabs.state.MinelabsProperties;
+import be.uantwerpen.minelabs.util.MinelabsProperties;
+import be.uantwerpen.minelabs.util.NucleusState;
+import be.uantwerpen.minelabs.util.NuclidesTable;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.client.MinecraftClient;
@@ -23,8 +25,6 @@ import net.minecraft.util.ActionResult;
 import net.minecraft.util.ItemScatterer;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
-import be.uantwerpen.minelabs.util.NuclidesTable;
-import be.uantwerpen.minelabs.util.NucleusState;
 import net.minecraft.util.math.ColorHelper;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
