+++ src/main/java/be/uantwerpen/minelabs/mixins/CampfireBlockEntityMixin.java
@@ -2,7 +2,7 @@ package be.uantwerpen.minelabs.mixins;
 
 import be.uantwerpen.minelabs.block.entity.ICampfireBlockEntity;
 import be.uantwerpen.minelabs.item.IFireReaction;
-import be.uantwerpen.minelabs.util.Properties;
+import be.uantwerpen.minelabs.util.MinelabsProperties;
 import be.uantwerpen.minelabs.util.Tags;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.CampfireBlockEntity;
@@ -26,16 +26,16 @@ public abstract class CampfireBlockEntityMixin implements ICampfireBlockEntity {
 
     @Inject(method = "litServerTick", at = @At(value = "HEAD"))
     private static void injectServerTick(World world, BlockPos pos, BlockState state, CampfireBlockEntity campfire, CallbackInfo ci) {
-        int color = state.get(Properties.FIRE_COLOR);
+        int color = state.get(MinelabsProperties.FIRE_COLOR);
         for (int i = 0; i < campfire.getItemsBeingCooked().size(); i++) {
             if (campfire.getItemsBeingCooked().get(i).isIn(Tags.Items.FIRE_CHANGER)) {
                 IFireReaction item = (IFireReaction) campfire.getItemsBeingCooked().get(i).getItem();
                 color = item.getFireColor();
             }
         }
-        if (color != state.get(Properties.FIRE_COLOR)) {
+        if (color != state.get(MinelabsProperties.FIRE_COLOR)) {
             //Change 3 -> 19 if you don't want observer updates
-            world.setBlockState(pos, state.with(Properties.FIRE_COLOR, color), 3);
+            world.setBlockState(pos, state.with(MinelabsProperties.FIRE_COLOR, color), 3);
             //campfire.setCachedState(state.with(Properties.FIRE_COLOR, color));
         }
     }
@@ -44,7 +44,7 @@ public abstract class CampfireBlockEntityMixin implements ICampfireBlockEntity {
     private static void craftServerTick(World world, BlockPos pos, BlockState state, CampfireBlockEntity campfire, CallbackInfo ci) {
         ICampfireBlockEntity castedE = (ICampfireBlockEntity) campfire;
         //Change 3 -> 19 if you don't want observer updates
-        world.setBlockState(pos, state.with(Properties.FIRE_COLOR, castedE.getLatestFire()), 3);
+        world.setBlockState(pos, state.with(MinelabsProperties.FIRE_COLOR, castedE.getLatestFire()), 3);
     }
 
     /**
