--- src/main/java/be/uantwerpen/minelabs/util/Properties.java
@@ -1,11 +1,0 @@
-package be.uantwerpen.minelabs.util;
-
-import net.minecraft.state.property.IntProperty;
-
-public class Properties {
-
-    public static final IntProperty FIRE_COLOR = IntProperty.of("fire_color", 0, 10);
-
-    public Properties() {
-    }
-}
